Refactor: Changed EngineTester's folder structure to match the current Engine project's folder structure. Added some new component folders in Engine.

+++ EngineTester/TestMain.cpp
@@ -1,7 +1,7 @@
 #include "gtest/gtest.h"
 #include "AllEngineTests.h"
 
-uint16 main(int argc, char* argv[])
+int main(int argc, char* argv[])
 {
 	::testing::InitGoogleTest(&argc, argv);
 	RUN_ALL_TESTS();
